refactor(status): centralize drift computation in handler
