+++ src/cli/commands/status.rs
@@ -1,9 +1,7 @@
 use anyhow::Context as _;
 
 use crate::config::TargetScope;
-use crate::deploy::TargetPath;
 use crate::engine::Engine;
-use crate::hash::sha256_hex;
 use crate::output::{JsonEnvelope, print_json};
 
 use super::Ctx;
@@ -16,32 +14,11 @@ struct NextActions {
 
 pub(crate) fn run(ctx: &Ctx<'_>, only: &[crate::cli::args::StatusOnly]) -> anyhow::Result<()> {
     #[derive(serde::Serialize)]
-    struct DriftItem {
-        target: String,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        root: Option<String>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        root_posix: Option<String>,
-        path: String,
-        path_posix: String,
-        expected: Option<String>,
-        actual: Option<String>,
-        kind: String,
-    }
-
-    #[derive(Default, serde::Serialize, Clone, Copy)]
-    struct DriftSummary {
-        modified: u64,
-        missing: u64,
-        extra: u64,
-    }
-
-    #[derive(serde::Serialize)]
     struct DriftSummaryByRoot {
         target: String,
         root: String,
         root_posix: String,
-        summary: DriftSummary,
+        summary: crate::handlers::status::DriftSummary,
     }
 
     #[derive(serde::Serialize)]
@@ -60,276 +37,24 @@ pub(crate) fn run(ctx: &Ctx<'_>, only: &[crate::cli::args::StatusOnly]) -> anyho
     warn_operator_assets_if_outdated(&engine, ctx.cli, &targets, &mut warnings, &mut next_actions)?;
     let prefix = action_prefix(ctx.cli);
 
-    let mut drift = Vec::new();
-    let mut summary = DriftSummary::default();
+    let report = crate::handlers::status::status_drift_report(
+        &desired,
+        &roots,
+        warnings,
+        crate::handlers::status::ExtraScanHashMode::IncludeHashes,
+    )?;
+    let warnings = report.warnings;
+    let drift = report.drift;
+    let summary = report.summary;
+    let any_manifest = report.any_manifest;
 
-    let mut manifests: Vec<Option<crate::target_manifest::TargetManifest>> =
-        vec![None; roots.len()];
-    let mut manifest_paths: Vec<Option<std::path::PathBuf>> = vec![None; roots.len()];
-    for (idx, root) in roots.iter().enumerate() {
-        let preferred = crate::target_manifest::manifest_path_for_target(&root.root, &root.target);
-        let legacy = crate::target_manifest::legacy_manifest_path(&root.root);
-
-        let (path, used_legacy) = if preferred.exists() {
-            (preferred, false)
-        } else if legacy.exists() {
-            (legacy, true)
-        } else {
-            continue;
-        };
-
-        if used_legacy {
-            warnings.push(format!(
-                "target manifest ({}): using legacy manifest filename {} (consider running `agentpack deploy --apply` to migrate)",
-                root.target,
-                path.display(),
-            ));
-        }
-
-        let (manifest, manifest_warnings) =
-            crate::target_manifest::read_target_manifest_soft(&path, &root.target);
-        warnings.extend(manifest_warnings);
-        manifests[idx] = manifest;
-        if manifests[idx].is_some() {
-            manifest_paths[idx] = Some(path);
-        }
-    }
-
-    let any_manifest = manifests.iter().any(Option::is_some);
-    if !any_manifest {
-        warnings.push(
-            "no target manifests found; drift may be inaccurate (run deploy --apply to write manifests)"
-                .to_string(),
-        );
+    if report.needs_deploy_apply {
         next_actions
             .human
             .insert(format!("{prefix} deploy --apply"));
         next_actions
             .json
             .insert(format!("{prefix} deploy --apply --yes --json"));
-        for (tp, desired_file) in &desired {
-            let expected = format!("sha256:{}", sha256_hex(&desired_file.bytes));
-            let root = super::super::util::best_root_idx(&roots, &tp.target, &tp.path)
-                .and_then(|idx| roots.get(idx))
-                .map(|r| r.root.as_path());
-            match std::fs::read(&tp.path) {
-                Ok(actual_bytes) => {
-                    let actual = format!("sha256:{}", sha256_hex(&actual_bytes));
-                    if actual != expected {
-                        summary.modified += 1;
-                        drift.push(DriftItem {
-                            target: tp.target.clone(),
-                            root: root.map(|p| p.to_string_lossy().to_string()),
-                            root_posix: root.map(crate::paths::path_to_posix_string),
-                            path: tp.path.to_string_lossy().to_string(),
-                            path_posix: crate::paths::path_to_posix_string(&tp.path),
-                            expected: Some(expected),
-                            actual: Some(actual),
-                            kind: "modified".to_string(),
-                        });
-                    }
-                }
-                Err(err) if err.kind() == std::io::ErrorKind::NotFound => {
-                    summary.missing += 1;
-                    drift.push(DriftItem {
-                        target: tp.target.clone(),
-                        root: root.map(|p| p.to_string_lossy().to_string()),
-                        root_posix: root.map(crate::paths::path_to_posix_string),
-                        path: tp.path.to_string_lossy().to_string(),
-                        path_posix: crate::paths::path_to_posix_string(&tp.path),
-                        expected: Some(expected),
-                        actual: None,
-                        kind: "missing".to_string(),
-                    });
-                }
-                Err(err) => return Err(err).context("read deployed file"),
-            }
-        }
-    } else {
-        let mut desired_by_root: Vec<Vec<(&TargetPath, &crate::deploy::DesiredFile)>> =
-            vec![Vec::new(); roots.len()];
-        for (tp, desired_file) in &desired {
-            let Some(root_idx) = super::super::util::best_root_idx(&roots, &tp.target, &tp.path)
-            else {
-                continue;
-            };
-            desired_by_root[root_idx].push((tp, desired_file));
-        }
-
-        for (idx, root) in roots.iter().enumerate() {
-            let Some(manifest) = &manifests[idx] else {
-                if desired_by_root[idx].is_empty() {
-                    continue;
-                }
-                warnings.push(format!(
-                    "no usable target manifest for {} {}; drift may be incomplete (run deploy --apply to write manifests)",
-                    root.target,
-                    root.root.display()
-                ));
-                next_actions
-                    .human
-                    .insert(format!("{prefix} deploy --apply"));
-                next_actions
-                    .json
-                    .insert(format!("{prefix} deploy --apply --yes --json"));
-                for (tp, desired_file) in &desired_by_root[idx] {
-                    let expected = format!("sha256:{}", sha256_hex(&desired_file.bytes));
-                    match std::fs::read(&tp.path) {
-                        Ok(actual_bytes) => {
-                            let actual = format!("sha256:{}", sha256_hex(&actual_bytes));
-                            if actual != expected {
-                                summary.modified += 1;
-                                drift.push(DriftItem {
-                                    target: tp.target.clone(),
-                                    root: Some(root.root.to_string_lossy().to_string()),
-                                    root_posix: Some(crate::paths::path_to_posix_string(
-                                        &root.root,
-                                    )),
-                                    path: tp.path.to_string_lossy().to_string(),
-                                    path_posix: crate::paths::path_to_posix_string(&tp.path),
-                                    expected: Some(expected),
-                                    actual: Some(actual),
-                                    kind: "modified".to_string(),
-                                });
-                            }
-                        }
-                        Err(err) if err.kind() == std::io::ErrorKind::NotFound => {
-                            summary.missing += 1;
-                            drift.push(DriftItem {
-                                target: tp.target.clone(),
-                                root: Some(root.root.to_string_lossy().to_string()),
-                                root_posix: Some(crate::paths::path_to_posix_string(&root.root)),
-                                path: tp.path.to_string_lossy().to_string(),
-                                path_posix: crate::paths::path_to_posix_string(&tp.path),
-                                expected: Some(expected),
-                                actual: None,
-                                kind: "missing".to_string(),
-                            });
-                        }
-                        Err(err) => return Err(err).context("read deployed file"),
-                    }
-                }
-                continue;
-            };
-
-            let mut managed_paths = crate::deploy::ManagedPaths::new();
-            for f in &manifest.managed_files {
-                let rel_path = std::path::Path::new(&f.path);
-                if rel_path.is_absolute()
-                    || rel_path
-                        .components()
-                        .any(|c| matches!(c, std::path::Component::ParentDir))
-                {
-                    warnings.push(format!(
-                        "target manifest ({}): skipped invalid entry path {:?} in {}",
-                        root.target,
-                        f.path,
-                        manifest_paths[idx]
-                            .as_ref()
-                            .map(|p| p.display().to_string())
-                            .unwrap_or_else(|| "<unknown>".to_string()),
-                    ));
-                    continue;
-                }
-
-                managed_paths.insert(TargetPath {
-                    target: root.target.clone(),
-                    path: root.root.join(&f.path),
-                });
-            }
-
-            for tp in &managed_paths {
-                let expected = desired
-                    .get(tp)
-                    .map(|f| format!("sha256:{}", sha256_hex(&f.bytes)));
-                match std::fs::read(&tp.path) {
-                    Ok(actual_bytes) => {
-                        let actual = format!("sha256:{}", sha256_hex(&actual_bytes));
-                        if let Some(exp) = &expected {
-                            if &actual != exp {
-                                summary.modified += 1;
-                                drift.push(DriftItem {
-                                    target: tp.target.clone(),
-                                    root: Some(root.root.to_string_lossy().to_string()),
-                                    root_posix: Some(crate::paths::path_to_posix_string(
-                                        &root.root,
-                                    )),
-                                    path: tp.path.to_string_lossy().to_string(),
-                                    path_posix: crate::paths::path_to_posix_string(&tp.path),
-                                    expected: Some(exp.clone()),
-                                    actual: Some(actual),
-                                    kind: "modified".to_string(),
-                                });
-                            }
-                        } else {
-                            summary.extra += 1;
-                            drift.push(DriftItem {
-                                target: tp.target.clone(),
-                                root: Some(root.root.to_string_lossy().to_string()),
-                                root_posix: Some(crate::paths::path_to_posix_string(&root.root)),
-                                path: tp.path.to_string_lossy().to_string(),
-                                path_posix: crate::paths::path_to_posix_string(&tp.path),
-                                expected: None,
-                                actual: Some(actual),
-                                kind: "extra".to_string(),
-                            });
-                        }
-                    }
-                    Err(err) if err.kind() == std::io::ErrorKind::NotFound => {
-                        if let Some(exp) = expected {
-                            summary.missing += 1;
-                            drift.push(DriftItem {
-                                target: tp.target.clone(),
-                                root: Some(root.root.to_string_lossy().to_string()),
-                                root_posix: Some(crate::paths::path_to_posix_string(&root.root)),
-                                path: tp.path.to_string_lossy().to_string(),
-                                path_posix: crate::paths::path_to_posix_string(&tp.path),
-                                expected: Some(exp),
-                                actual: None,
-                                kind: "missing".to_string(),
-                            });
-                        }
-                    }
-                    Err(err) => return Err(err).context("read deployed file"),
-                }
-            }
-
-            if !root.scan_extras {
-                continue;
-            }
-            if !root.root.exists() {
-                continue;
-            }
-
-            let mut files = crate::fs::list_files(&root.root)?;
-            files.sort();
-            for path in files {
-                if crate::target_manifest::is_target_manifest_path(&path) {
-                    continue;
-                }
-
-                let tp = TargetPath {
-                    target: root.target.clone(),
-                    path: path.clone(),
-                };
-                if managed_paths.contains(&tp) {
-                    continue;
-                }
-
-                summary.extra += 1;
-
-                drift.push(DriftItem {
-                    target: tp.target.clone(),
-                    root: Some(root.root.to_string_lossy().to_string()),
-                    root_posix: Some(crate::paths::path_to_posix_string(&root.root)),
-                    path: tp.path.to_string_lossy().to_string(),
-                    path_posix: crate::paths::path_to_posix_string(&tp.path),
-                    expected: None,
-                    actual: Some(format!("sha256:{}", sha256_hex(&std::fs::read(&tp.path)?))),
-                    kind: "extra".to_string(),
-                });
-            }
-        }
     }
 
     if summary.modified > 0 || summary.missing > 0 {
@@ -378,12 +103,12 @@ pub(crate) fn run(ctx: &Ctx<'_>, only: &[crate::cli::args::StatusOnly]) -> anyho
     let (mut drift, summary, summary_total_opt) = if only_kinds.is_empty() {
         (drift, summary_total, None)
     } else {
-        let drift: Vec<DriftItem> = drift
+        let drift: Vec<crate::handlers::status::DriftItem> = drift
             .into_iter()
             .filter(|d| only_kinds.contains(d.kind.as_str()))
             .collect();
 
-        let mut summary = DriftSummary::default();
+        let mut summary = crate::handlers::status::DriftSummary::default();
         for d in &drift {
             match d.kind.as_str() {
                 "modified" => summary.modified += 1,
@@ -396,28 +121,29 @@ pub(crate) fn run(ctx: &Ctx<'_>, only: &[crate::cli::args::StatusOnly]) -> anyho
         (drift, summary, Some(summary_total))
     };
 
-    let summary_by_root = |drift: &[DriftItem]| -> Vec<DriftSummaryByRoot> {
-        let mut by_root: std::collections::BTreeMap<(String, String), DriftSummaryByRoot> =
-            std::collections::BTreeMap::new();
-        for d in drift {
-            let root = d.root.as_deref().unwrap_or("<unknown>").to_string();
-            let root_posix = d.root_posix.as_deref().unwrap_or("<unknown>").to_string();
-            let key = (d.target.clone(), root_posix.clone());
-            let entry = by_root.entry(key).or_insert_with(|| DriftSummaryByRoot {
-                target: d.target.clone(),
-                root,
-                root_posix,
-                summary: DriftSummary::default(),
-            });
-            match d.kind.as_str() {
-                "modified" => entry.summary.modified += 1,
-                "missing" => entry.summary.missing += 1,
-                "extra" => entry.summary.extra += 1,
-                _ => {}
+    let summary_by_root =
+        |drift: &[crate::handlers::status::DriftItem]| -> Vec<DriftSummaryByRoot> {
+            let mut by_root: std::collections::BTreeMap<(String, String), DriftSummaryByRoot> =
+                std::collections::BTreeMap::new();
+            for d in drift {
+                let root = d.root.as_deref().unwrap_or("<unknown>").to_string();
+                let root_posix = d.root_posix.as_deref().unwrap_or("<unknown>").to_string();
+                let key = (d.target.clone(), root_posix.clone());
+                let entry = by_root.entry(key).or_insert_with(|| DriftSummaryByRoot {
+                    target: d.target.clone(),
+                    root,
+                    root_posix,
+                    summary: crate::handlers::status::DriftSummary::default(),
+                });
+                match d.kind.as_str() {
+                    "modified" => entry.summary.modified += 1,
+                    "missing" => entry.summary.missing += 1,
+                    "extra" => entry.summary.extra += 1,
+                    _ => {}
+                }
             }
-        }
-        by_root.into_values().collect()
-    };
+            by_root.into_values().collect()
+        };
     let summary_by_root = summary_by_root(&drift);
 
     if ctx.cli.json {
@@ -520,7 +246,10 @@ pub(crate) fn run(ctx: &Ctx<'_>, only: &[crate::cli::args::StatusOnly]) -> anyho
                     b.path.as_str(),
                 ))
         });
-        let by_root: std::collections::BTreeMap<(String, String), DriftSummary> = summary_by_root
+        let by_root: std::collections::BTreeMap<
+            (String, String),
+            crate::handlers::status::DriftSummary,
+        > = summary_by_root
             .into_iter()
             .map(|s| ((s.target, s.root), s.summary))
             .collect();
