+++ src/handlers/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod deploy;
 pub(crate) mod evolve;
 pub(crate) mod read_only;
 pub(crate) mod rollback;
+pub(crate) mod status;
